Started cleaning up components and linking new structure
