+++ src/servers/main.rs
@@ -1,4 +1,4 @@
-use crate::{constants::MAIN_PORT, show_error, TARGET};
+use crate::{api::TARGET, constants::MAIN_PORT, ui::show_error};
 use reqwest::{
     header::{self, HeaderMap, HeaderValue},
     Client,
